Switch to data URIs for Firefox 51+
Since https://bugzil.la/898728, Firefox unloads Object URLs on window
unload, so we can no longer create Object URLs for a document we haven't
loaded yet. Technically, we could create Object URLs for resources
requested after the document has loaded, but for simplicity we don't (we
can't just create data URIs in the bootstrapper, because we need
extended base64 support to create them, which we have in core.js).

Unscientific testing shows not much of a performance hit.

Also drop support for Firefox <= 34 in the process, which was released
late 2014 and the first ESR *after* which is already no longer
supported.