[laskyawm] Contain resizing to workspace on top and left sides
Also, snap windows to workspace instead of window

Both are achieved by containing all windows in a #windows element with
the dimensions of the workspace. This also allows us to position windows
relative to the workspace instead of the screen, and refactor
forceMinimize() and others to work with those numbers.