Switch back to Object URLs in Chrome
For compatibility with importing in Firetext and Strut.

Since Chrome 54, Object URLs being considered insecure caused FileReader
to stop working.

It turns out Object URLs are only considered insecure in Chrome *inside
Data URLs*. So we switch the top-level html resource and all
subresources to Object URLs.

Since 5830338, we always used Data URLs for html resources, for
compatibility with aprendio-html. That app is currently broken for
unrelated reasons. However, when working around that, this commit does
indeed regress it: scripts are blocked as insecure after navigating,
even though there are no Data URLs in play. This is worrying, but a
working File API is more important.

Since 00e289f, we always used Data URLs for all resources in Chrome in
secure contexts, because Chrome seemed to consider Object URLs insecure.